+++ src/Products/PluginIndexes/unindex.py
@@ -54,6 +54,7 @@ class UnIndex(SimpleItem):
     """Simple forward and reverse index.
     """
 
+    zmi_icon = 'icon-info'
     _counter = None
     operators = ('or', 'and')
     useOperator = 'or'