+++ src/Products/ZCatalog/ZCatalog.py
@@ -99,6 +99,7 @@ class ZCatalog(Folder, Persistent, Implicit):
     security.declareProtected(search_zcatalog, 'all_meta_types')
 
     meta_type = 'ZCatalog'
+    zmi_icon = 'icon-search'
 
     manage_options = (
         {'label': 'Contents', 'action': 'manage_main'},