+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 - Fix a TypeError on Python 3 when trying to lookup in an OOBTree
   a value for a key that has an invalid type
 
+- Add icon for Bootstrap ZMI.
 
 4.1 (2018-03-06)
 ----------------