+++ CHANGES.rst
@@ -155,10 +155,10 @@ Changelog
 ------------------
 
 - Fix BooleanIndex when index inversion occurs as a result of reindexing
-  and existing document with the opposite value. Fixes LP #1236354.
+  and existing document with the opposite value. Fixes LPÂ #1236354.
 
 3.0 (2013-02-24)
-----------------
+------------------
 
 - Strip white space from name when adding a column or index.