+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 ----------------
 
 - Fix a TypeError on Python 3 when trying to lookup in an OOBTree
-  a value for a key that has an invalid type
+  a value for a key that has an invalid type.
+  (`#36 <https://github.com/zopefoundation/Products.ZCatalog/pull/36>`_)
 
 
 4.1 (2018-03-06)